Support for ds-identify
This patch adds support for using the datasource with cloud-init
installations that leverage ds-identify
(https://github.com/number5/cloud-init/blob/master/tools/ds-identify).

The ds-identify program uses bash types (functions, scripts, etc.) to
determine whether or not a configured datasource is valid on the
installed system. The script in this patch allows the ds-identify
program to query whether or not the VMwareGuestInfo datasource is valid.
Otherwise, depending on how ds-identify is configured, a system
configured to use the VMwareGuestInfo datasource might actually have it
disabled after ds-identify fails to query the datasource's availability.