Add ability to clear guestinfo keys after reading
This patch adds the ability to clear data from the guestinfo dictionary
after the data has been read by the datasource. This feature is enabled
by adding the following to the metadata:

        cleanup-guestinfo:
        - userdata
        - vendordata

When the above snippet is added to a VM's metadata, the keys
"guestinfo.userdata", "guestinfo.userdata.encoding",
"guestinfo.vendordata", and "guestinfo.vendordata.encoding" will all be
set to the value of a single space character.