Get subp object based on cloud-init version
cloud-init >= 20.3, subp is a seperate module.
So, data source should be modified accordingly. If someoone is using
guestinfo data source with cloud-init <=20.2 we should not break things.

See the below link for more details:
https://github.com/canonical/cloud-init/commit/3c551f6ebc12f7729a2755c89b19b9000e27cc88

This patch takes care of the above things.
Tested on Photon 3.0 VM with Photon OS using package test scripts.

Signed-off-by: Shreenidhi Shedi <sshedi@vmware.com>