Remove deepmerge dependency
This patch removes the deepmerge module as a required dependency
of this cloud-init datasource. There is still an option to use
the deepmerge module if the environment variable
CLOUD_INIT_VMWARE_GUEST_INFO_MERGE_STRATEGY is set to 'deepmerge'.
Otherwise a deep merge is performed with a recursive strategy using
Python built-ins.